requirements.txt
Django==1.7.4
Markdown==2.5.2
django-appconf==0.6
django-ckeditor==4.4.7
django-compressor==1.4
django-filter==0.9.2
django-tinymce==1.5.3
django-wysiwyg==0.7.0
djangorestframework==3.0.4
drf-nested-routers==0.9.0
six==1.9.0
wsgiref==0.1.2
